fixes wrapValueBasedOnType not actually returning a provider when asked for it
