+++ src/main/groovy/com/wooga/gradle/test/IntegrationSpec.groovy
@@ -63,91 +63,7 @@ class IntegrationSpec extends nebula.test.IntegrationSpec {
 
     // TODO: To be deprecated in the future by a better implementation
     String wrapValueBasedOnType(Object rawValue, String type, Closure<String> fallback = null) {
-        def value
-
-        def match = StringTypeMatch.match(type)
-        type = match.mainType
-        def subType = match.subType
-
-        switch (type) {
-            case "Closure":
-                def returnType = subType ?: rawValue.class.typeName
-                value = "{${wrapValueBasedOnType(rawValue, returnType, fallback)}}"
-                break
-                // TODO:
-            case "Callable":
-                def returnType = subType ?: rawValue.class.typeName
-                value = "new java.util.concurrent.Callable<${returnType}>() {@Override ${returnType} call() throws Exception {${wrapValueBasedOnType(rawValue, returnType, fallback)}}"
-                break
-            case "Object":
-                value = "new Object() {@Override String toString() { ${wrapValueBasedOnType(rawValue, "String", fallback)} }}"
-                break
-            case "Directory":
-                value = "project.layout.projectDirectory.dir(${wrapValueBasedOnType(rawValue, String, fallback)})"
-                break
-            case "RegularFile":
-                value = "project.layout.projectDirectory.file(${wrapValueBasedOnType(rawValue, String, fallback)})"
-                break
-            case "Provider":
-                switch (subType) {
-                    case "RegularFile":
-                        value = "project.layout.buildDirectory.file(${wrapValueBasedOnType(rawValue, String, fallback)})"
-                        break
-                    case "Directory":
-                        value = "project.layout.buildDirectory.dir(${wrapValueBasedOnType(rawValue, String, fallback)})"
-                        break
-                    default:
-                        value = "project.provider(${wrapValueBasedOnType(rawValue, "Closure<${subType}>", fallback)})"
-                        break
-                }
-                break
-                // TODO: ...
-            case "java.lang.String":
-            case "String":
-                value = "'${escapedPath(rawValue.toString())}'"
-                break
-                // TODO: Assumes that the raw value is a collection, no auto-conversion to collection
-            case "[]":
-                value = "${wrapValueBasedOnType(rawValue, "List<${subType}>", fallback)} as ${subType}[]"
-                break
-                // TODO: Assumes that the raw value is a collection, no auto-conversion to collection
-            case "...":
-                value = "${rawValue.collect { wrapValueBasedOnType(it, subType, fallback) }.join(", ")}"
-                break
-            case "File":
-                value = "new File(${wrapValueBasedOnType(rawValue, String.class, fallback)})"
-                break
-                // TODO: Assumes that the raw value is a collection, no auto-conversion to collection
-            case "java.util.ArrayList":
-            case "List":
-                def returnType = subType ?: String.class.typeName
-                value = "[${rawValue.collect { wrapValueBasedOnType(it, returnType, fallback) }.join(", ")}]"
-                break
-                // TODO: Assumes that the raw value is a collection, no auto-conversion to collection
-            case "Map":
-                if (subType) {
-                    def split = subType.split(",", 2).collect { it.trim() }
-                    def keyType = split.first()
-                    def valueType = split.last()
-                    value = "[" + rawValue.collect { k, v -> "${wrapValueBasedOnType(k, keyType, fallback)} : ${wrapValueBasedOnType(v, valueType, fallback)}" }.join(", ") + "]"
-                } else {
-                    value = "[" + rawValue.collect { k, v -> "${wrapValueBasedOnType(k, k.getClass(), fallback)} : ${wrapValueBasedOnType(v, v.getClass(), fallback)}" }.join(", ") + "]"
-                }
-                value = value == "[]" ? "[:]" : value
-                break
-            default:
-                if (fallback) {
-                    fallback.setDelegate(this)
-                    if (fallback.getMaximumNumberOfParameters() == 1) {
-                        value = fallback(type)
-                    } else if (fallback.getMaximumNumberOfParameters() == 3) {
-                        value = fallback(rawValue, type, fallback)
-                    }
-                } else {
-                    value = rawValue
-                }
-        }
-        value
+        return PropertyUtils.wrapValueBasedOnType(rawValue, type, fallback)
     }
 
     static File generateBatchWrapper(String fileName, Boolean printEnvironment = false) {
